Update plots to include 0 score
Vibe coded using GPT-5 mini. Here is the prompt:

This is a tool to produce a report with grading results for university
assignments. It includes an abundance of tables and plots to help us
visualize the grading results in different sections of the same course.
As of now, in the code, there are 6 possible score values: 1, 2, 3, 4,
5. However, we have just updated our rubric, and there are now 6
possible score values: 0, 1, 2, 3, 4, 5. This program needs to be
updated to include the new level of 0. In the visualizations, the 0
should be colored dark grey.